Add files via upload
The project involves using tools and algorithms that help in identifying and correcting invalid, incomplete, or inconsistent data. The techniques used are tailored based on the specific data and business requirements to ensure that the data is reliable and can be used to its full potential. In summary, this project seeks to optimize the value of data by making it high-quality and trustworthy for further processing and analysis.